add some engine code

+++ src/arena.rs
@@ -17,8 +17,8 @@ impl Handle {
     }
 
     /// Returns the index of this handle.
-    pub const fn index(self) -> u32 {
-        self.0
+    pub const fn index(self) -> usize {
+        self.0 as usize
     }
 }
